requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
openapi-codec==1.3.2
pdfminer3k==1.3.4
ply==3.11
protobuf==3.17.3
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
slate3k==0.5.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
